implemented tustin's namebox

+++ include/graphics.h
@@ -3,7 +3,7 @@
 
 // overworld
 extern const u8 gSignpostWindow_Gfx[];
-extern const u8 gMessageBox_Gfx[];
+extern const u32 gMessageBox_Gfx[];
 extern const u16 gMessageBox_Pal[];
 
 // pokeballs
